Engraving: move easeinout.h/cpp from dom to compat/midi
