+++ src/engraving/compat/midi/easeinout.h
